move utils in a folder
