+++ torchtitan/experiments/vlm/datasets/utils/image.py
