+++ torchtitan/experiments/vlm/datasets/utils/packing.py
